Final changes.

+++ src/com/company/Record.java
@@ -3,9 +3,9 @@ package com.company;
 public class Record {
 
     public String name;
-    public Integer record;
+    public long record;
 
-    public Record(String name, Integer record) {
+    public Record(String name, long record) {
         this.name = name;
         this.record = record;
     }
